Notice board api ipml

+++ src/main/java/com/pplip/domain/board/notice/usecase/impl/NoticeCommentServiceImpl.java
@@ -1,6 +1,7 @@
 package com.pplip.domain.board.notice.usecase.impl;
 
 import com.pplip.domain.auth.persistence.entity.Account;
+import com.pplip.domain.auth.utils.SecurityUtils;
 import com.pplip.domain.board.notice.api.request.NoticeCommentRequest;
 import com.pplip.domain.board.notice.api.response.NoticeCommentResponse;
 import com.pplip.domain.board.notice.persistence.dao.NoticeCommentDao;
@@ -18,6 +19,8 @@ import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.List;
+
 @RequiredArgsConstructor
 @Service
 @Transactional
@@ -28,6 +31,16 @@ public class NoticeCommentServiceImpl implements NoticeCommentService {
     private final NoticeCommentParser parser;
     @Override
     public Page<NoticeCommentResponse.Summary> findAllByBoardId(Long boardId, PageRequest pageRequest) {
+        List<NoticeCommentResponse.Summary> datas = dao.findAll(boardId, pageRequest);
+
+        if (!SecurityUtils.isAnonymous()) {
+            Long userId = SecurityUtils.getCurrentUser().getUserId();
+            datas = datas.stream().map(data -> {
+                data.setAuthor(data.getAuthorId() == userId);
+                return data;
+            }).toList();
+        }
+
         return new Page<>(dao.findAll(boardId, pageRequest),
                 pageRequest.getPageNum(),
                 pageRequest.getPageSize(),
@@ -43,7 +56,11 @@ public class NoticeCommentServiceImpl implements NoticeCommentService {
         if (dao.insert(entity) != 1) {
             throw new BoardLogicException(ErrorCode.COMMENT_CREATE_ERROR);
         }
-        return dao.findById(entity.getId()).orElseThrow(() -> new BoardLogicException(ErrorCode.COMMENT_NOT_FOUND));
+        NoticeCommentResponse.Detail resData = dao.findById(entity.getId())
+                .orElseThrow(() -> new BoardLogicException(ErrorCode.COMMENT_NOT_FOUND));
+
+        resData.setAuthor(resData.getAuthorId() == userId);
+        return resData;
     }
 
     @Override
@@ -52,19 +69,36 @@ public class NoticeCommentServiceImpl implements NoticeCommentService {
                                                UserDetails userDetails) {
         Long userId = ((Account) userDetails).getUserId();
 
-        NoticeCommentModel model = new NoticeCommentModel(update, id, userId);
-        dao.update(model.toEntity());
+        NoticeComment entity = dao.findByIdToEntity(id)
+                .orElseThrow(() -> new BoardLogicException(ErrorCode.COMMENT_NOT_FOUND, "댓글을 찾을 수 없습니다."));
+
+        if (entity.getAuthorId() != userId) {
+            throw new BoardLogicException(ErrorCode.FORBIDDEN, "작성자만 수정 가능합니다.");
+        }
+
+        entity.updateContent(update.getContent());
+        dao.update(entity);
 
         NoticeCommentResponse.Detail detail = dao.findById(id)
                 .orElseThrow(() -> new BoardLogicException(ErrorCode.COMMENT_NOT_FOUND));
+
+        detail.setAuthor(detail.getAuthorId() == userId);
 
         return parser.detailResToUpdateRes(detail);
     }
 
     @Override
-    public void delete(Long id) {
+    public void delete(Long id, UserDetails userDetails) {
+        Long userId = ((Account) userDetails).getUserId();
+        NoticeComment entity = dao.findByIdToEntity(id)
+                .orElseThrow(() -> new BoardLogicException(ErrorCode.COMMENT_NOT_FOUND, "댓글을 찾지 못하였습니다."));
+
+        if (entity.getAuthorId() != userId) {
+            throw new BoardLogicException(ErrorCode.FORBIDDEN, "작성자만 삭제할 수 있습니다.");
+        }
+
         if (dao.delete(id) != 1) {
-            throw new BoardLogicException(ErrorCode.COMMENT_DELETE_FAIL_ERROR);
+            throw new BoardLogicException(ErrorCode.COMMENT_DELETE_FAIL_ERROR, "댓글 삭제에 실패했습니다.");
         }
     }
 }
